fix(aws_handler): fix aws handler timeouts
Fix usage of 'proxy' method in aws-serverless-express. Follow usage examples that intentionally drop the 'callback' argument of the serverless handler. Explicitly invoke 'proxy' rather than bind arguments that result in invalid arguments sent to the proxy function.

aws-serverless-express changed the signature of the 'proxy' function, introducing a 'resolutionMode' argument in advance of the 'callback' (https://github.com/awslabs/aws-serverless-express/blob/master/src/index.js#L197). When arguments from the serverless handler are automatically propagated to the 'proxy' function, this results in 'aws-serverless-express' assuming (but not validating) that the lambda callback is a defined 'resolutionMode'. Ultimately, it causes the lambda to never resolve and timeout. An alternative might be to bind and pass an explicit 'resolutionMode' argument, but it's not document by 'aws-serverless-express' so we should avoid it. Another alternative might be to pin this library to <3.3.0 of 'aws-serverless-express' (when this was not a problem), but this change should be preferred as its backwards compatible.